requirements.txt
openai==1.3.7
python-dotenv==1.0.0
python-telegram-bot==20.7